[DEV-10593] organization and importing styles
