+++ packages/chart/src/scss/main.scss
@@ -60,7 +60,6 @@
 
 .cdc-open-viz-module.type-chart {
   @import 'DataTable';
-  @import '../components/Legend/LegendGroup/index.scss';
 
   .legend-top {
     .legend-wrapper .tooltip-boundary {
