+++ packages/map/index.html
@@ -19,7 +19,7 @@
   <body>
     <!-- DEFAULT EXAMPLES -->
     <!-- <div class="react-container" data-config="/examples/private/nhis.json"></div> -->
-    <div class="react-container" data-config="/examples/hex-colors.json"></div>
+    <!-- <div class="react-container" data-config="/examples/hex-colors.json"></div> -->
     <!-- <div class="react-container react-container--maps" data-config="/examples/annotation/index.json">/</div> -->
 
     <!-- <div class="react-container react-container--maps" data-config="/examples/private/tooltip-issue.json"></div> -->
@@ -50,7 +50,7 @@
     <!-- <div class="react-container react-container--maps" data-config="/examples/hex-with-arrows.json"></div> -->
 
     <!-- TP4 EXAMPLES -->
-    <!-- <div class="react-container react-container&#45;&#45;maps" data-config="/examples/example-city-state.json"></div> -->
+    <div class="react-container react-container&#45;&#45;maps" data-config="/examples/example-city-state.json"></div>
     <!-- <div
       class="react-container react-container--maps"
       data-config="/examples/example-city-state-no-territories.json"
